Using anonymous unions and structs to emulate the 16 bit register mechanic
Gameboys have the ability to pair certain 8 bit registers to form 16 bit
registers. By using unions and structs we can simulate this behavior in
an elegant way in c++.

We can access any base register as expected like `my_registers.a` or
`my_registers.b`. But we can also access our 16 bit registers in the
same way: `my_registers.ab`. This should make it easier to work with
registers.